Ability to specify logdir
nginx_log_dir points to /var/log/nginx, but sometimes it is
required that logs be stored on a different FS.

If nginx_log_symlink_src is specified, then the directory will
be created and the symlink /var/log/nginx will point to the
new location.

If the directory exists and contains any data, it will be renamed
to /var/log/nginx.bkp previously to force the symlink creation.